PHASE_DFLASH: switch gates to fp16 — sm_75 has no BF16 support
First Gate-3.5 run failed at vLLM engine init with:
  ValueError: Bfloat16 is only supported on GPUs with compute
  capability of at least 8.0. Your Quadro RTX 6000 GPU has
  compute capability 7.5. You can use float16 instead by
  explicitly setting the `dtype` flag in CLI, for example:
  --dtype=half.

Both target weights (Qwen/Qwen3.6-27B, BF16 safetensors) and
drafter weights (z-lab/Qwen3.6-27B-DFlash, BF16 safetensors) are
trained in BF16. vLLM casts BF16 -> fp16 at load time.

Numerical implications for our gates:
  - Gate-3.5 (determinism, np=1 vs np=2 byte-identical): unchanged
    by the dtype switch. The test is internal consistency across
    batch shapes; fp16 cast is identical across both shapes.
  - Gate 0 (speedup repro): now reports fp16 numbers, not BF16.
    Speedup ratio is the gate's metric, not absolute throughput,
    so the comparison stays valid. Absolute throughput may differ
    from z-lab's published number (which was on A100/H100 BF16).

Risk note for fp16: Qwen3.6 27B has rope_theta=10M and trained
n_ctx=262144. Long-context fp16 can overflow activations in
extreme cases. Our determinism harness uses prompts < 50 tokens
and max_tokens=96 → well clear of the fp16 overflow regime.

+++ scripts/gate0-dflash-speedup.py
@@ -76,7 +76,7 @@ def run_pass(spec_enabled: bool, label: str) -> dict:
     init_kwargs = dict(
         model=TARGET_DIR,
         tensor_parallel_size=2,
-        dtype="bfloat16",
+        dtype="float16",             # sm_75 has no BF16
         gpu_memory_utilization=0.92,
         cpu_offload_gb=12,
         enforce_eager=True,
